+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Support Token base security authentication
 - Support for Scheduled Autoscaling
 - Support for `dbVersion` field in Autonomous databases Container database resources
+- Support for Archive Log Backup and Point in time restore
 
 ## 3.80.0 (June 10, 2020)